+++ lib/rubocop-rspec.rb
@@ -2,14 +2,11 @@
 
 require 'rubocop'
 
-# TODO: workaround for RuboCop renaming
-Rubocop = RuboCop
-
 require 'rubocop/rspec/version'
 require 'rubocop/rspec/inject'
 require 'rubocop/rspec/top_level_describe'
 
-Rubocop::RSpec::Inject.defaults!
+RuboCop::RSpec::Inject.defaults!
 
 # cops
 require 'rubocop/cop/rspec/describe_class'