+++ lib/rubocop/cop/rspec/example_wording.rb
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-module Rubocop
+module RuboCop
   module Cop
     module RSpec
       # Do not use should when describing your tests.