+++ lib/rubocop/cop/rspec/instance_variable.rb
@@ -1,6 +1,6 @@
 # encoding: utf-8
 
-module Rubocop
+module RuboCop
   module Cop
     module RSpec
       # When you have to assign a variable instead of using an instance