+++ spec/project_spec.rb
@@ -16,7 +16,7 @@ describe 'RuboCop Project' do # rubocop:disable RSpec/DescribeClass
     end
 
     subject(:default_config) do
-      Rubocop::ConfigLoader.load_file('config/default.yml')
+      RuboCop::ConfigLoader.load_file('config/default.yml')
     end
 
     it 'has configuration for all cops' do