+++ spec/rubocop/cop/rspec/describe_method_spec.rb
@@ -2,7 +2,7 @@
 
 require 'spec_helper'
 
-describe Rubocop::Cop::RSpec::DescribeMethod do
+describe RuboCop::Cop::RSpec::DescribeMethod do
   subject(:cop) { described_class.new }
 
   it 'enforces non-method names' do